+++ src/main/java/com/netflix/nfgraph/compressed/NFCompressedGraphPointers.java
@@ -24,42 +24,42 @@ import com.netflix.nfgraph.exception.NFGraphException;
 
 /**
  * This class holds all of the offsets into the {@link NFCompressedGraph}'s byte array.<p/>
- * 
+ *
  * This class maintains a mapping of type name to int array.  For a given type, the offset in the {@link NFCompressedGraph}'s byte array
  * where the connections for a given node are encoded is equal to the value of the int array for the node's type at the index for the node's ordinal.<p/>
- * 
+ *
  * It is unlikely that this class will need to be used externally.
  */
 public class NFCompressedGraphPointers {
 
-    private final Map<String, int[]>pointersByOrdinal;
-    
+    private final Map<String, long[]>pointersByOrdinal;
+
     public NFCompressedGraphPointers() {
-        this.pointersByOrdinal = new HashMap<String, int[]>();
+        this.pointersByOrdinal = new HashMap<String, long[]>();
     }
-    
+
     /**
-     * @return the offset into the {@link NFCompressedGraph}'s byte array for the node identified by the given type and ordinal. 
+     * @return the offset into the {@link NFCompressedGraph}'s byte array for the node identified by the given type and ordinal.
      */
-    public int getPointer(String nodeType, int ordinal) {
-        int pointers[] = pointersByOrdinal.get(nodeType);
+    public long getPointer(String nodeType, int ordinal) {
+        long pointers[] = pointersByOrdinal.get(nodeType);
         if(pointers == null)
             throw new NFGraphException("Undefined node type: " + nodeType);
         if(ordinal < pointers.length)
             return pointers[ordinal];
         return -1;
     }
-    
-    public void addPointers(String nodeType, int pointers[]) {
+
+    public void addPointers(String nodeType, long pointers[]) {
         pointersByOrdinal.put(nodeType, pointers);
     }
-    
+
     public int numPointers(String nodeType) {
         return pointersByOrdinal.get(nodeType).length;
     }
-    
-    public Map<String, int[]> asMap() {
+
+    public Map<String, long[]> asMap() {
         return pointersByOrdinal;
     }
-    
+
 }
